backend/requirements.txt
Django==3.2.2
requests==2.25.1
django-filter==2.4.0
djangorestframework==3.12.4
django-rest-auth==0.9.5
# Swagger:
drf-yasg==1.21.0
# Adaptador de la base de datos relacional PostgreSQL para Python.
psycopg2-binary>=2.8
# Toolkit que permite aprovechar al máximo el uso interactivo de Python.
# Muy útil cuando utilizemos el shell de Django.
ipython==8.8.0
django-cors-headers
pytube
spotipy
beautifulsoup4
requests
pandas
gunicorn==20.1.0
whitenoise
djoser
djangorestframework-simplejwt
pyjwt
django-environ
